Ensure index.html will be regenerated after switching branch
After creating a new .deb file and switching to the gh-pages branch, the
index.html will be checked out with a newer timestamp than the .deb
file. That means the index.html target is newer than its prerequisites.

With this change an intermediate index.html.tmp file is generated every
time, and index.html is updated if it differs.