Exclude shared libraries from auto-provides processing
This ensures that the gcc-latest package isn't installed in order to
satisfy rpm-build-libs' dependency on libgomp.so.1()(64bit) during the
build. The libraries provided by gcc-latest should never be used to
satisfy dependencies for other packages, because the gcc-latest
libraries are not installed in global directories searched by the
dynamic loader.

This conforms to the Fedora Packaging Guidelines section
https://docs.fedoraproject.org/en-US/packaging-guidelines/AutoProvidesAndRequiresFiltering/

  "Packages must not provide RPM dependency information when that
  information is not global in nature, or are otherwise handled (e.g.
  through a virtual provides system). e.g. a plugin package containing a
  binary shared library must not "provide" that library unless it is
  accessible through the system library paths."

Also disable libssp and libffi.