Add postint and postrm scripts for deb package
We need to run mkheaders after installing the deb, in case the 16.04
headers do not match the ones on the machine doing the install. That
also means we need to remove them on uninstall, so they aren't left
hanging around.

This also adds install and remove tests to the container build, just
to check the scripts aren't completely broken.

Fixes #6